+++ packages/nitro-basic/src/functions/uiCreateClientApp.ts
@@ -1,7 +1,9 @@
 import type { App } from 'vue'
 import type { Router } from 'vue-router'
+import { dxtFunctionalPlugin } from '@dxtmisha/functional'
 
 import { initClientRouter } from './initClientRouter'
+import type { NitroAppOptions } from '../types/nitroAppTypes'
 
 /**
  * Finalizes the initialization and mount of the Vue application on the client side.
@@ -15,14 +17,18 @@ import { initClientRouter } from './initClientRouter'
  * @param app current Vue application instance / текущий экземпляр приложения Vue
  * @param rootContainer selector or element for mounting / селектор или элемент для монтирования
  * @param router instance of Vue Router / экземпляр Vue Router
+ * @param options configuration options for the application / параметры конфигурации приложения
  * @param action additional action to perform before mounting / дополнительное действие перед монтированием
  */
 export async function uiCreateClientApp<T>(
   app: App<T>,
   rootContainer: string | T = '#app',
   router?: Router | undefined,
+  options: NitroAppOptions = {},
   action?: (app: App<T>) => Promise<void> | void
 ) {
+  app.use(dxtFunctionalPlugin, options)
+
   await initClientRouter(router)
 
   if (action) {