+++ packages/nitro-basic/src/functions/uiCreateServerApp.ts
@@ -2,6 +2,7 @@ import type { App } from 'vue'
 import type { Router } from 'vue-router'
 import type { SSRContext } from 'vue/server-renderer'
 import { Geo, MetaStatic } from '@dxtmisha/functional-basic'
+import { dxtFunctionalPlugin } from '@dxtmisha/functional'
 
 import { useHeaders } from '../composables/useHeaders'
 
@@ -13,6 +14,8 @@ import { initServerStorage } from './initServerStorage'
 import { initScriptsJson } from './initScriptsJson'
 import { initSsrApp } from './initSsrApp'
 
+import type { NitroAppOptions } from '../types/nitroAppTypes'
+
 /**
  * Initializes the server-side application, including storage, routing, and SSR rendering.
  *
@@ -20,6 +23,7 @@ import { initSsrApp } from './initSsrApp'
  * @param app root component of the application / корневой компонент приложения
  * @param request incoming server request / входящий запрос сервера
  * @param router optional router instance / экземпляр роутера (опционально)
+ * @param options configuration options for the application / параметры конфигурации приложения
  * @param action additional action to perform before rendering / дополнительное действие перед рендерингом
  * @param context SSR context for the renderer / контекст SSR для рендерера
  * @param body HTML template for substitution / HTML-шаблон для подстановки
@@ -29,6 +33,7 @@ export async function uiCreateServerApp<T>(
   app: App<T>,
   request: Request,
   router?: Router | undefined,
+  options: NitroAppOptions = {},
   action?: (app: App<T>) => Promise<void> | void,
   context: SSRContext = {},
   body?: string
@@ -37,7 +42,10 @@ export async function uiCreateServerApp<T>(
   initServerStorage(app)
   initCookieStorage(app, request)
 
-  app.runWithContext(() => initApi(request))
+  app.runWithContext(() => {
+    initApi(request)
+    app.use(dxtFunctionalPlugin, options)
+  })
 
   await initServerRouter(request, router)