+++ packages/nitro-basic/src/library.ts
@@ -2,7 +2,6 @@
 export * from './composables/useHeaders'
 
 // Functions
-export * from './functions/getInject'
 export * from './functions/getRequestOrigin'
 export * from './functions/getRequestUrl'
 export * from './functions/initApi'
@@ -13,6 +12,7 @@ export * from './functions/initScriptsJson'
 export * from './functions/initServerRouter'
 export * from './functions/initServerStorage'
 export * from './functions/initSsrApp'
+export * from './functions/uiBootstrapClient'
 export * from './functions/uiBootstrapServer'
 export * from './functions/uiCookieStorage'
 export * from './functions/uiCreateApp'